+++ config/nodemailer-config.js
@@ -1,0 +1,13 @@
+const nodemailer = require("nodemailer");
+
+require("dotenv").config();
+
+const transporter = nodemailer.createTransport({
+  service: "gmail",
+  auth: {
+    user: process.env.SMTP_USER,
+    pass: process.env.SMTP_PASS,
+  },
+});
+
+module.exports = transporter;
