+++ helpers/mailing-helper.js
@@ -1,13 +1,18 @@
-const nodemailer = require("nodemailer");
+const transporter = require("../config/nodemailer-config");
 
-require("dotenv").config();
-
-const transporter = nodemailer.createTransport({
-  service: "gmail",
-  auth: {
-    user: process.env.SMTP_USER,
-    pass: process.env.SMTP_PASS,
-  },
-});
-
-module.exports = transporter;
+exports.sendCreatePasswordMail = async (student) => {
+  try {
+    await transporter.sendMail({
+      from: "animosys022@gmail.com",
+      to: student.email,
+      subject: "Animo.sys - Finish your account registration",
+      html: `
+      Hello ${student.name.first}, <br>
+      To fully verify your Animo.sys account, please create your desired password
+      through this <a href=${"/"}>link</a>.
+      `,
+    });
+  } catch (error) {
+    console.log(error);
+  }
+};
